requirements.txt
keras~=2.12.0
opencv-python~=4.7.0.72
numpy~=1.23.5
tensorflow~=2.12.0
pandas~=2.0.2
scikit-learn~=1.2.2
python-telegram-bot==20.4